Abstract test framework integration out of appmap.pytest
